moved all typing to separate module
