+++ asyncstdlib/contextlib.py
@@ -15,7 +15,7 @@ from collections import deque
 from functools import partial
 import sys
 
-from ._typing import Protocol, AsyncContextManager, ContextManager
+from ._typing import Protocol, AsyncContextManager, ContextManager, T
 from ._core import awaitify
 from ._utility import public_module, slot_get as _slot_get
 
@@ -30,7 +30,6 @@ class ACloseable(Protocol):
         """Asynchronously close this object"""
 
 
-T = TypeVar("T")
 AC = TypeVar("AC", bound=ACloseable)
